Added error handling.
Added error handling (SIError)

Removed the ability to delete materials.

Due to an issue with deleting materials causing fatal errors (see issues page on GitHub for more) I commented out this function.

error handling with alerts is implemented in *most* places (you will still see some where it is not). Because this uses the LocalizedError type in Swift, it's implemented quite cleanly in the views.

**COMING SOON**

I want deleting functionality to still exist, but it needs to be handled properly. Implementing more safe guards on what can and cannot be deleted is coming, and hopefully, more information for the user.

I want to redo the readme. The current simple checklist is fine, but adding images and/or videos to show off the app's functionality is coming soon.

Solving the color changing issue. I realized why it is happening, just need to figure out how to deal with it moving forward.

Still working to understand photos for products, but that function, as well as changing colors hopefully, is on the horizon.

Since the app is beginning to grow, implementing MVVM principles is a must to clean up the code on the views.